Fix paths. Correct named and default exports.
